rddt_clone/requirements.txt
Django==2.2
django-crispy-forms==1.7.2
django-discover-runner==1.0
django-filter==2.1.0
django-role-permissions==2.2.1
Pillow==6.0.0
pytz==2018.9
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
